DEP: Fixed Seaborn Deprecated Functions
Changed factorplot functions to the new catplot functions. Changed the tsplot function to the lineplot function. Replaced the size attribute by the new height attribute in catplot. The only visual change in the generated graph, is that instead of graphing all training/validation loss curves in one plot, it now plots only the average and uses a band to reflect the spread.